+++ commerce/payment/src/main/java/ru/yandex/practicum/service/PaymentService.java
@@ -3,14 +3,15 @@ package ru.yandex.practicum.service;
 import ru.yandex.practicum.dto.OrderDto;
 import ru.yandex.practicum.dto.PaymentDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface PaymentService {
     PaymentDto payment(OrderDto orderDto);
 
-    Double productCost(OrderDto orderDto);
+    BigDecimal productCost(OrderDto orderDto);
 
-    Double getTotalCost(OrderDto orderDto);
+    BigDecimal getTotalCost(OrderDto orderDto);
 
     void paymentSuccess(UUID paymentId);
